==> CONTENEDOR-PYSPARK/01-ETL-tipo_gasto.ipynb <==
#### NOTA: En este cuaderno se creara las Tablas `tipo_gasto` del Datawarehouse

***
### Importando Librerias

from pyspark.sql import SparkSession

***
Creando Session de Spark

spark = SparkSession.builder \
    .appName('tipo_gasto') \
    .config('spark.ui.enabled', 'true') \
    .getOrCreate()

***
Importando Conexion a Contenedor `mod4_mysql`

from db.connection import connec_db

***
Trabajando La Tabla `tipo_gasto`

tipo_gasto = spark.read.csv("Data/TiposDeGasto.csv", sep=",", header=True, inferSchema=True)

***
Visualizando como vienen los Datos

tipo_gasto.show()

***
Visualizando la Inferencia del Schema

tipo_gasto.printSchema()

***
Cambio Nombre de las Columnas

tipo_gasto = tipo_gasto.withColumnRenamed("Descripcion", "TipoGasto")
tipo_gasto = tipo_gasto.withColumnRenamed("Monto_Aproximado", "MontoAproximado")
tipo_gasto.show()

***
Creando la Conexion

engine = connec_db()

***
Pasando el Dataframe a un Pandas.Dataframe  

# IMPORTANDO LIBRERIA PANDAS
import pandas as pd

pandas_df = tipo_gasto.toPandas()

pandas_df.to_csv("csv_practica/tipo_gasto.csv", index=False)

***
Vaciando el Dataframe a la Tabla `tipo_gasto`

pandas_df.to_sql(name="tipo_gasto", con=engine, index=False, if_exists="append")

spark.stop()
