==> CONTENEDOR-PYSPARK/03-ETL-producto-tipo_producto.ipynb <==
#### NOTA: En este cuaderno se crearan las Tablas `producto` y `tipo_producto` del Datawarehouse

***
### Importando Librerias

from pyspark.sql import SparkSession

from pyspark.sql.functions import  count, when, isnull,  monotonically_increasing_id


***
Creando Session de Spark

spark = SparkSession.builder.appName('producto-tipo_producto').getOrCreate()

***
Importando Conexion a Contenedor `mod4_mysql`

from db.connection import connec_db

***
Trabajando La Tabla `producto`

producto = spark.read.csv("Data/PRODUCTOS.csv", sep=",", header=True, inferSchema=True)

***
Visualizando como vienen los Datos

producto.show(5)

***
Visualizando la Inferencia del Schema

producto.printSchema()

***
Anomalia en Registros de Tabla `producto`
<image src="assets/ANOMALIA-MAESTRO-PRODUCTOS.png">

***
Verficar en ventas si esos Productos han tenido ventas y si sus valores se ven mas normales

venta = spark.read.csv("Data/Venta.csv", sep=",", header=True, inferSchema=True )
venta.show(1)

FILTRANDO LAS VENTAS CUYOS CODIGOS DE PRODUCTOS SEAN LOS QUE TIENEN ANOMALIAS`

codigos_anomalias = [42802, 42803, 42804, 42805, 42806, 42807, 42808, 42809]
venta.filter(venta.IdProducto.isin(codigos_anomalias)).show()

Al no tener ventas estos Productos con Fuertes Anomalias y sin Descripcion puedo eliminarlos

del venta

producto.filter(producto.ID_PRODUCTO.isin(codigos_anomalias)).show()

producto = producto.filter(~producto.ID_PRODUCTO.isin(codigos_anomalias))

producto.filter(producto.ID_PRODUCTO.isin(codigos_anomalias)).show()

del codigos_anomalias

***
Buscando Valores Faltantes

# Contar valores nulos por columna
null_counts = producto.agg(*[count(when(isnull(c), c)).alias(c) for c in producto.columns])

null_counts.show()

*** 
Reemplazando Valores Faltantes

# REEMPLAZA LOS VALORES FALTANTES DE LA COLUMNA TIPO CON LA PALABRA `VARIOS`
producto = producto.fillna({'Tipo': 'VARIOS'})

# VERIFICO QUE NO QUEDO NULOS
null_counts = producto.agg(*[count(when(isnull(c), c)).alias(c) for c in producto.columns]).show()

del null_counts

Creo el Dataframe tipo_producto que luego será vaciado en la Tabla `tipo_producto`

tipo_producto = producto.select(producto['Tipo']).distinct()

# Agregar una columna con un ID único e incremental le sumo 1 para comenzar desde 1 y no de 0
tipo_producto = tipo_producto.withColumn('IdTipoProducto', monotonically_increasing_id()+1)

# Renombro columna
tipo_producto = tipo_producto.withColumnRenamed('Tipo', 'TipoProducto')
tipo_producto = tipo_producto.select('IdTipoProducto', 'TipoProducto')
tipo_producto.show()

Ahora necesito crear la Columna IdTipoProducto en `producto`, mapeando lo que tenga en "Tipo" <br>
Para mapear se trabaja Tipo SQL se crea un Join de Tablas

producto_join_tipo = producto.join(tipo_producto, producto.Tipo == tipo_producto.TipoProducto)

producto = producto_join_tipo.select('ID_PRODUCTO', 'Concepto', 'Precio', 'IdTipoProducto')

producto.columns

***
Cambio Nombre de las Columnas

producto = producto.withColumnRenamed("ID_PRODUCTO", "IdProducto")
producto = producto.withColumnRenamed("Concepto", "Producto")
producto.show(5)

***
Creando la Conexion

engine = connec_db()

***
Pasando los Dataframes a Dataframes de Pandas  

# IMPORTANDO LIBRERIA PANDAS
import pandas as pd

pandas_df_producto = producto.toPandas()
pandas_df_tipo_producto = tipo_producto.toPandas()

Creando Csv para practica

pandas_df_tipo_producto.to_csv("csv_practica/tipo_producto.csv", index=False)

pandas_df_producto.to_csv("csv_practica/producto.csv", index=False)

***
Vaciando el Dataframe a la Tabla `tipo_gasto`

pandas_df_tipo_producto.to_sql(name="tipo_producto", con=engine, index=False, if_exists="append")

pandas_df_producto.to_sql(name="producto", con=engine, index=False, if_exists="append")

spark.stop()
