==> CONTENEDOR-PYSPARK/04-ETL-localidad-provincia.ipynb <==
#### NOTA: En este cuaderno se crearan las Tablas `localidad`, `provincia`, <br> `proveedor`, `cliente` y `sucursal` del Datawarehouse

***
### Importando Librerias

from pyspark.sql import SparkSession
from pyspark.sql.types import IntegerType, FloatType

from pyspark.sql.functions import count, when, isnull, monotonically_increasing_id, initcap
from pyspark.sql.functions import initcap, when, col, regexp_replace, coalesce, avg

***
Creando Session de Spark

spark = SparkSession.builder.appName('localidad-provincia-etc').getOrCreate()

***
Importando Conexion a Contenedor `mod4_mysql`

from db.connection import connec_db

***
Trabajando Las Tablas `proveedor`, `sucursal` y `cliente` 

proveedor = spark.read.csv("Data/Proveedores.csv", sep=",", header=True, inferSchema=True)

sucursal = spark.read.csv("Data/Sucursales.csv", sep=";", header=True, inferSchema=True)

cliente = spark.read.csv("Data/Clientes.csv", sep=";", header=True, inferSchema=True)

***
Visualizando como vienen los Datos

proveedor.show(5)

sucursal.show(5)


cliente.show(5)

***
Visualizando la Inferencia de los Schemas

proveedor.printSchema()

sucursal.printSchema()

cliente.printSchema()

***
Llevando Campos a Letra Capital

# TABLA cliente
cliente = cliente.withColumn('Localidad', initcap('Localidad'))
cliente = cliente.withColumn('Provincia', initcap('Provincia'))
cliente = cliente.withColumn('Domicilio', initcap('Domicilio'))
cliente = cliente.withColumn('Nombre_y_Apellido', initcap('Nombre_y_Apellido'))

# TABLA sucursal
sucursal = sucursal.withColumn('Localidad', initcap('Localidad'))
sucursal = sucursal.withColumn('Provincia', initcap('Provincia'))
sucursal = sucursal.withColumn('Direccion', initcap('Direccion'))

# TABLA proveedor
proveedor = proveedor.withColumn('Address', initcap('Address'))
proveedor = proveedor.withColumn('City', initcap('City'))
proveedor = proveedor.withColumn('State', initcap('State'))
proveedor = proveedor.withColumn('Country', initcap('Country'))
proveedor = proveedor.withColumn('departamen', initcap('departamen'))

cliente.show(1)
sucursal.show(1)
proveedor.show(1)

***
Normalizacion de "Buenos Aires", "Córdoba" y "Capital Federal" y "Río Negro" en los campos  `Localidad` y `Provincia` Tomar en cuenta que en la Tabla proveedor esos campos todavía se llaman `City` y `State`

reemplazar_1 = [
    "Bs.as. ",
    "B.aires",
    "B. Aires",
    "B.Aires",
    " Bs.As.",
    "Bs.As. ",
    "Bs As",
    "Bs.As.",
    "Buenos Aires",
    "C Debuenos Aires",
    "Caba",
    "Cdad De Buenos Aires",
    "Ciudad de Buenos Aires",
    "Ciudad De Buenos Aires",
    "Cdad de Buenos Aires",
    "Pcia Bs AS",
    "Pcia Bs As",
    "Prov de Bs As.",
    "Prov De Bs As.",
    "Prov de Buenos Aires.",
    "Provincia de Buenos Aires",
    "Provincia De Buenos Aires",
    "CABA",
    "C deBuenos Aires"]
# Se reemplazaran por "Buenos Aires"

reemplazar_2 = ["Coroba", "Cordoba",  "Cã³rdoba"]
# Se reemplazaran por "Córdoba"

reemplazar_3 = [
    "Boca De Atencion Monte Castro",
    "Cap.   Federal",
    "Cap. Fed.",
    "Capfed",
    "CapFed",
    "Cap. Fed.",
    "Cap. Federal",
    "Capital"
]
# Se reemplazaran por "Capital Federal"

reemplazar_4 = ["Rio Negro"]
# Se reemplazara por "Río Negro"

Normalizando Buenos Aires

# TABLA `sucursal`
sucursal = sucursal.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_1), 'Buenos Aires')\
                               .otherwise(col('Localidad')))

sucursal = sucursal.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_1), 'Buenos Aires')\
                               .otherwise(col('Provincia')))

# TABLA `cliente`
cliente = cliente.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_1), 'Buenos Aires')\
                               .otherwise(col('Localidad')))

cliente = cliente.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_1), 'Buenos Aires')\
                               .otherwise(col('Provincia')))

# TABLA `proveedor`
proveedor = proveedor.withColumn('City', when(col('City')\
                               .isin(reemplazar_1), 'Buenos Aires')\
                               .otherwise(col('City')))

proveedor = proveedor.withColumn('State', when(col('State')\
                               .isin(reemplazar_1), 'Buenos Aires')\
                               .otherwise(col('State')))


Normalizando Córdoba


# TABLA `sucursal`
sucursal = sucursal.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_2), 'Córdoba')\
                               .otherwise(col('Localidad')))

sucursal = sucursal.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_2), 'Córdoba')\
                               .otherwise(col('Provincia')))

# TABLA `cliente`
cliente = cliente.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_2), 'Córdoba')\
                               .otherwise(col('Localidad')))

cliente = cliente.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_2), 'Córdoba')\
                               .otherwise(col('Provincia')))

# TABLA `proveedor`
proveedor = proveedor.withColumn('City', when(col('City')\
                               .isin(reemplazar_2), 'Córdoba')\
                               .otherwise(col('City')))

proveedor = proveedor.withColumn('State', when(col('State')\
                               .isin(reemplazar_2), 'Córdoba')\
                               .otherwise(col('State')))

Normalizando Capital Federal

# TABLA `sucursal`
sucursal = sucursal.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_3), 'Capital Federal')\
                               .otherwise(col('Localidad')))

sucursal = sucursal.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_3), 'Capital Federal')\
                               .otherwise(col('Provincia')))

# TABLA `cliente`
cliente = cliente.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_3), 'Capital Federal')\
                               .otherwise(col('Localidad')))

cliente = cliente.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_3), 'Capital Federal')\
                               .otherwise(col('Provincia')))

# TABLA `proveedor`
proveedor = proveedor.withColumn('City', when(col('City')\
                               .isin(reemplazar_3), 'Capital Federal')\
                               .otherwise(col('City')))

proveedor = proveedor.withColumn('State', when(col('State')\
                               .isin(reemplazar_3), 'Capital Federal')\
                               .otherwise(col('State')))

Normalizando Río Negro

# TABLA `sucursal`
sucursal = sucursal.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_4), 'Río Negro')\
                               .otherwise(col('Localidad')))

sucursal = sucursal.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_4), 'Río Negro')\
                               .otherwise(col('Provincia')))

# TABLA `cliente`
cliente = cliente.withColumn('Localidad', when(col('Localidad')\
                               .isin(reemplazar_4), 'Río Negro')\
                               .otherwise(col('Localidad')))

cliente = cliente.withColumn('Provincia', when(col('Provincia')\
                               .isin(reemplazar_4), 'Río Negro')\
                               .otherwise(col('Provincia')))

# TABLA `proveedor`
proveedor = proveedor.withColumn('City', when(col('City')\
                               .isin(reemplazar_4), 'Río Negro')\
                               .otherwise(col('City')))

proveedor = proveedor.withColumn('State', when(col('State')\
                               .isin(reemplazar_4), 'Río Negro')\
                               .otherwise(col('State')))

del reemplazar_1
del reemplazar_2
del reemplazar_3
del reemplazar_4

cliente.show(1)
sucursal.show(1)
proveedor.show(1)

***
Normalizando Nombre de Sucursales

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Mendoza1"]), 'Mendoza-1')\
                               .otherwise(col('Sucursal')))

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Mendoza2"]), 'Mendoza-2')\
                               .otherwise(col('Sucursal')))

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Palermo 1"]), 'Palermo-1')\
                               .otherwise(col('Sucursal')))

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Palermo 2"]), 'Palermo-2')\
                               .otherwise(col('Sucursal')))

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["MDQ1"]), 'Mdq-1')\
                               .otherwise(col('Sucursal')))

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["MDQ2"]), 'Mdq-2')\
                               .otherwise(col('Sucursal')))

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Rosario1"]), 'Rosario-1')\
                               .otherwise(col('Sucursal')))

sucursal = sucursal.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Rosario2"]), 'Rosario-2')\
                               .otherwise(col('Sucursal')))

***
Corrigiendo `Latitud` y `Longitud` reemplazando la `,` por `.`

# TABLA sucursal
sucursal = sucursal.withColumn('Latitud', regexp_replace('Latitud', ',', '.'))
sucursal = sucursal.withColumn('Longitud', regexp_replace('Longitud', ',', '.'))
# TABLA cliente
cliente = cliente.withColumn('X', regexp_replace('X', ',', '.'))
cliente = cliente.withColumn('Y', regexp_replace('Y', ',', '.'))

Corrigiendo `Latitud` y `Longitud` casteando columnas al tipo float

# TABLA sucursal
sucursal = sucursal.withColumn('Latitud', sucursal['Latitud'].cast(FloatType()))
sucursal = sucursal.withColumn('Longitud', sucursal['Longitud'].cast(FloatType()))
# TABLA cliente
cliente = cliente.withColumn('X', cliente['X'].cast(FloatType()))
cliente = cliente.withColumn('Y', cliente['Y'].cast(FloatType()))

cliente.show(1)
sucursal.show(1)
proveedor.show(1)

***
Buscando Valores Faltantes

proveedor.agg(*[count(when(isnull(c), c)).alias(c) for c in proveedor.columns]).show()
cliente.agg(*[count(when(isnull(c), c)).alias(c) for c in cliente.columns]).show()
sucursal.agg(*[count(when(isnull(c), c)).alias(c) for c in sucursal.columns]).show()


***
Llenando Campos Nulos con "Sin Dato" o 0 segun sea el caso

# TABLA cliente
cliente = cliente.fillna({'Nombre_y_Apellido':'Sin Dato'})
cliente = cliente.fillna({'Domicilio':'Sin Dato'})
cliente = cliente.fillna({'Telefono':'Sin Dato'})
cliente = cliente.fillna({'X':0})
cliente = cliente.fillna({'Y':0})
# TABLA proveedor
proveedor = proveedor.fillna({'Nombre':'Sin Dato'})

***
Filtrar en `cliente` los registros cuya "Localidad" o "Provincia" son Nulas

filtrados = cliente.filter(col('Localidad').isNull() | col('Provincia').isNull())
filtrados.show(32)

Dado que los registros cuya Localidad y/o Provincia que presentaron valores Nulos, tenían Latitud y Longitud, se buscaron aparte los Nombres de las Localidades y Provincias a las cuales pertenecian

provincias_faltantes = [    
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Buenos Aires',
 'Córdoba',
 'Córdoba',
 'Córdoba',
 'Córdoba',
 'Córdoba',
 'Neuquén',
 'Neuquén',
 'Tucumán',
 'Tucumán',
 'Tucumán',
 'Tucumán'
]


localidades_faltantes = [
    'Aeropuerto Internacional Ezeiza',
    'Ángel Etcheverry',
    'Berisso',
    'Guernica',
    'Guernica',
    'La Plata',
    'La Plata',
    'Abasto',
    'La Plata',
    'Adrogué',
    'Adrogué',
    'Adrogué',
    'Pilar Sur',
    'El Cazador',
    'Ruta Nacional 8',
    'Jardín de Infantes 918',
    'Tercera Sección',
    'San Fernando',
    'Ruta Nacional 8',
    'El Peligro',
    'Ruta Nacional 8',
    'Peña',
    'Mariano Mansilla',
    'José Esteban Echeverría',
    'Hugo Juárez',
    'Ruta Nacional 8',
    'San Martín de los Andes',
    'Huilqui Menuco',
    'Alederetes',
    'Tafí Viejo',
    'El Manantial',
    'Santa Bárbara'
]

Creo un Dataframe con los Valores Faltantes y le creo un indice autoincremental

import numpy as np

Se crea un Dataframe con las provincias y localidades faltantes con un indice de 0 a 31

#Index que voy a poner a coincidir con filtrados
indice = np.arange(32).astype(IntegerType) # el indice va de 0 a 31
valores_faltantes_df = spark.createDataFrame(zip(indice, provincias_faltantes, localidades_faltantes),
                                              schema=['index','prov_falta', 'loc_falta'])

del indice

Se le crea a filtrado el indice de 0 a 31 para que coincida con valores_faltantes_df

filtrados = filtrados.withColumn('index', monotonically_increasing_id())

Se unen ambos Dataframes a traves de la columna index y se borra dicha columna

filtrados = filtrados.join(valores_faltantes_df, on='index').drop('index')

filtrados.show(32)

Ya filtrados tiene los valores faltantes... <br>
Se crea un join con cliente por medio del campo ID

# filtrados = filtrados.drop(filtrados.Localidad).drop(filtrados.Provincia)
filtra2 = filtrados.select("ID",'prov_falta', 'loc_falta')
filtra2 = filtra2.withColumnRenamed("ID", "id-f")

# Realizar un left join entre 'cliente' y 'filtrados'
cliente_join_left = cliente.join(filtra2, cliente['ID'] == filtra2['id-f'], 'left')

# Actualizar las columnas 'Localidad' y 'Provincia' con los valores faltantes
cliente_join_left = cliente_join_left.withColumn('Localidad', coalesce(filtra2['loc_falta'], 
                                                                       cliente['Localidad']))

cliente_join_left = cliente_join_left.withColumn('Provincia', coalesce(filtra2['prov_falta'], 
                                                                       cliente['Provincia']))

# Eliminar las columnas adicionales de 'filtrados' si es necesario
cliente_actualizado = cliente_join_left.drop(filtra2['loc_falta']).drop(filtra2['prov_falta']).drop(filtra2['id-f'])

cliente = cliente_actualizado

del filtra2
del filtrados
del cliente_join_left
del cliente_actualizado
del valores_faltantes_df
del provincias_faltantes
del localidades_faltantes

cliente.show()

***
Se crea un Rango Etario

cliente = cliente.withColumn(
    'RangoEtario',
    when(cliente['Edad'] <= 30, '1_Hasta 30 años')
    .when((cliente['Edad'] > 30) & (cliente['Edad'] <= 40), '2_De 31 a 40 años')
    .when((cliente['Edad'] > 40) & (cliente['Edad'] <= 50), '3_De 41 a 50 años')
    .when((cliente['Edad'] > 50) & (cliente['Edad'] <= 60), '4_De 51 a 60 años')
    .otherwise('5_Desde 61 años')
)

Borro Columna col10

cliente = cliente.drop(cliente.col10)

cliente.show(1)

***
Empezando a Trabajar la Tabla localidad

# Se crean Todas las combinaciones existentes en las distintas tablas (Pero sin duplicados)
temp_cliente_localidad = cliente.select("Localidad", "Provincia").dropDuplicates()
temp_sucursal_localidad = sucursal.select("Localidad", "Provincia").dropDuplicates()
temp_proveedor_localidad = proveedor.select("City", "State").dropDuplicates()

temp_proveedor_localidad = temp_proveedor_localidad.withColumnRenamed("City", "Localidad")\
                                                   .withColumnRenamed("State", "Provincia")

# Concateno todo en localidad usando union
localidad = temp_cliente_localidad\
            .union(temp_sucursal_localidad)\
            .union(temp_proveedor_localidad)

# Elimino Duplicados
localidad = localidad.dropDuplicates()

del temp_cliente_localidad
del temp_proveedor_localidad
del temp_sucursal_localidad

localidad.show()

Se Crea la tabla provincia

provincia = localidad.select("Provincia").dropDuplicates()
provincia = provincia.withColumn('IdProvincia', monotonically_increasing_id()+1)

provincia.show(1)

Se termina de crear la tabla localidad

localidad = localidad.withColumn('IdLocalidad', monotonically_increasing_id()+1)
localidad.show(truncate=False)

# Necesito utilizar una tabla con nombre del campo 'Provincia' diferente a la otra tabla
provincia_map = provincia.withColumnRenamed("Provincia", "Provi")

# Realizo el Join de las tablas localidad y provincia (en este caso representada por provincia.map)
localidad_join_provincia = localidad.join(provincia_map, localidad.Provincia == provincia_map.Provi)

# De todo el Join Selecciono las columnas que necesito
localidad = localidad_join_provincia.select("IdLocalidad", "Localidad", "IdProvincia")

# Creo una tabla para Mapear contra `cliente`, `proveedor`, `sucursal`
mapeo = localidad_join_provincia.select("IdLocalidad", "Provincia", "Localidad")
mapeo = mapeo.withColumnRenamed("Provincia", "Provi")\
             .withColumnRenamed("Localidad", "Locali")

del localidad_join_provincia

localidad.show(1)

Mapeo a la tabla `sucursal` con "mapeo" para colocarle un IdLocalidad que represente su <br>
combinacion `Localidad - Provincia`

sucursal_join_mapeo = sucursal.join(
    mapeo,
    (mapeo['Provi'] == sucursal['Provincia']) & (mapeo['Locali'] == sucursal['Localidad'])
)
sucursal_join_mapeo.show(1)

Selecciono los campos que necesito del Join <br>
Renombro la columna 'ID' a 'IdSucursal'

# Seleccionando los campos que necesito
sucursal = sucursal_join_mapeo.select("ID", "Sucursal", "Direccion", "IdLocalidad", "Latitud", "Longitud")

# Renombrando el Campo "ID"
sucursal = sucursal.withColumnRenamed("ID", "IdSucursal")

sucursal = sucursal.withColumnRenamed("Direccion", "Domicilio")

del sucursal_join_mapeo

sucursal.show()

Mapeo a la tabla `cliente` con "mapeo" para colocarle un IdLocalidad que represente su <br>
combinacion `Localidad - Provincia`

cliente_join_mapeo = cliente.join(
    mapeo,
    (mapeo['Provi'] == cliente['Provincia']) & (mapeo['Locali'] == cliente['Localidad'])
)
cliente_join_mapeo.show(1)

#Borro los campos que sobran
cliente_join_mapeo = cliente_join_mapeo.drop("Provincia","Localidad","Provi","Locali")

#Renombro Campos
cliente_join_mapeo = cliente_join_mapeo.withColumnRenamed("ID", "IdCliente")\
                                       .withColumnRenamed("X", "Longitud")\
                                       .withColumnRenamed("Y", "Latitud")\
                                       .withColumnRenamed("Nombre_y_Apellido", "NombreApellido")

cliente_join_mapeo.show()

cliente = cliente_join_mapeo

cliente = (cliente.withColumnRenamed("Usuario_Alta", "UsuarioAlta")\
                 .withColumnRenamed("Fecha_Alta", "FechaAlta")\
                 .withColumnRenamed("Fecha_Ultima_Modificacion", "FechaUpdate")\
                 .withColumnRenamed("Usuario_Ultima_Modificacion", "UsuarioUpdate")\
                 .withColumnRenamed("Marca_baja", "MarcaBaja"))


del cliente_join_mapeo

cliente.show(5)

Mapeo a la tabla `proveedor` con "mapeo" para colocarle un IdLocalidad que represente su <br>
combinacion `Localidad - Provincia`

proveedor_join_mapeo = proveedor.join(
    mapeo,
    (mapeo['Provi'] == proveedor['State']) & (mapeo['Locali'] == proveedor['City'])
)
proveedor_join_mapeo.show(1)

# Borro los campos que sobran
proveedor_join_mapeo = proveedor_join_mapeo.drop("City", "State", "Provi", "Locali")

# Renombro Campos
proveedor_join_mapeo = (proveedor_join_mapeo.withColumnRenamed("IDProveedor", "IdProveedor")
                                             .withColumnRenamed("Address", "Domicilio")
                                             .withColumnRenamed("Country", "Pais")
                                             .withColumnRenamed("departamen", "Departamento"))

proveedor_join_mapeo.show(1)

proveedor = proveedor_join_mapeo

del proveedor_join_mapeo

del provincia_map
del mapeo

proveedor.show(1)

*** 
Toca Imputar valores a las Longitudes y Latitudes NULAS a la cuales ya se les asigno 0 para identificarlas <br>
El acuerdo en clase fue Segun su localidad, ubicar el resto de clientes que tuvieran esa misma localidad y obtener el promedio de las latitudes y longitudes

Necesito Sacar un Nuevo Df a partir de Cliente Solo con los campos 'IdCliente' y 'Latitud' siempre y cuando Latitud sea diferente de cero(0) 

latitudes = cliente.filter(cliente['Latitud'] != 0).select('IdLocalidad', 'Latitud')

# Saco el Promedio de las latitudes por cada IdLocalidad
promedio_latitudes = latitudes.groupBy('IdLocalidad').agg(avg('Latitud').alias('LatitudPromedio'))
promedio_latitudes.show(1)

# Hago un Join left para tener en una columna la 'latitudPromedio'
cliente = cliente.join(promedio_latitudes, 'IdLocalidad', 'left')

# Mapeo las 'Laltitud` == 0 y hago que sean igaul a `LatitudPromedio`
cliente = cliente.withColumn('Latitud', when(col('Latitud') == 0, 
                             col('LatitudPromedio')).otherwise(col('Latitud')))
cliente.show(1)

# Elimino la Columna 'LatitudPromedio``
cliente = cliente.drop('LatitudPromedio')

del latitudes
del promedio_latitudes

Repito el Proceso pero ahora con 'Longitud'

longitudes = cliente.filter(cliente['Longitud'] != 0).select('IdLocalidad', 'Longitud')

promedio_longitudes = longitudes.groupBy('IdLocalidad').agg(avg('longitud').alias('longitudPromedio'))
promedio_longitudes.show(1)

# Hago un Join left para tener en una columna la 'longitudPromedio'
cliente = cliente.join(promedio_longitudes, 'IdLocalidad', 'left')

# Mapeo las 'Longitud` == 0 y hago que sean igaul a `LongitudPromedio`
cliente = cliente.withColumn('Longitud', when(col('Longitud') == 0, 
                             col('LongitudPromedio')).otherwise(col('Longitud')))
cliente.show(1)

# Elimino la Columna 'LongitudPromedio``
cliente = cliente.drop('LongitudPromedio')
cliente.show(1)

cliente.filter(cliente['Longitud'] == 0).show()

cliente.filter(cliente['Latitud'] == 0).show()

Tomar en cuenta que hay 3 clientes Los cuales son Unicos en sus Localidades por lo que es imposible calcular Promedio de otros clientes

cliente.agg(*[count(when(isnull(c), c)).alias(c) for c in cliente.columns]).show()

latitudes_nulas = cliente.filter(col('Latitud').isNull()).select('IdLocalidad')
latitudes_nulas.show()

longitudes_nulas = cliente.filter(col('longitud').isNull()).select('IdLocalidad')
longitudes_nulas.show()

localidad.filter(col("IdLocalidad") == 263).show()

localidad.filter(col("IdLocalidad") == 420).show()

localidad.filter(col("IdLocalidad") == 511).show()

provincia.filter(col("IdProvincia") == 6).show()

del longitudes_nulas
del latitudes_nulas

Se busco por internet manualmente la Longitud y Latitud Promedio de esas 3 Localidades

+ cliente con "IdLocalidad" == 420
    + Angel Etcheverry
    + "Latitud" = -34.56403
    + "Longitud" = -58.53444

+ cliente "IdLocalidad" == 263
    + Villa Brown
    + "Latitud"  = -35.0272
    + "Longitud" = -58.0719

+ cliente con "IdLocalidad" == 511
    + Parque San Lorenzo
    + "Latitud" = -34.87164
    + "Longitud" = -58.28075

# Cliente con IdLocalidad = 420
cliente = cliente.withColumn('Latitud', when(col('IdLocalidad') == 420, -34.56403).otherwise(col('Latitud')))
cliente = cliente.withColumn('Longitud', when(col('IdLocalidad') == 420, -58.53444).otherwise(col('Longitud')))

# Cliente con IdLocalidad = 263
cliente = cliente.withColumn('Latitud', when(col('IdLocalidad') == 263, -35.0272).otherwise(col('Latitud')))
cliente = cliente.withColumn('Longitud', when(col('IdLocalidad') == 263, -58.0719).otherwise(col('Longitud')))

# Cliente con IdLocalidad = 511
cliente = cliente.withColumn('Latitud', when(col('IdLocalidad') == 511, -34.87164).otherwise(col('Latitud')))
cliente = cliente.withColumn('Longitud', when(col('IdLocalidad') == 511, -58.28075).otherwise(col('Longitud')))

cliente.show(1)

***
Creando la Conexion

engine = connec_db()

***
Pasando los Dataframes a Dataframes de Pandas de a uno para evitar colapso

# IMPORTANDO LIBRERIA PANDAS
import pandas as pd

### Pasando Los Dataframes a Pandas para poder pasarlos a Tablas SQL

***
Vaciando el Dataframe a la Tabla `provincia`

pandas_df_provincia = provincia.toPandas()

pandas_df_provincia.to_sql(name="provincia", con=engine, index=False, if_exists="append")

pandas_df_provincia.to_csv("csv_practica/provincia.csv", index=False)

del pandas_df_provincia
del provincia

***
Vaciando el Dataframe a la Tabla `localidad`

pandas_df_localidad = localidad.toPandas()

pandas_df_localidad.to_sql(name="localidad", con=engine, index=False, if_exists="append")

pandas_df_localidad.to_csv("csv_practica/localidad.csv", index=False)

del localidad
del pandas_df_localidad

***
Vaciando el Dataframe a la Tabla `sucursal`

pandas_df_sucursal = sucursal.toPandas()

pandas_df_sucursal.to_sql(name="sucursal", con=engine, index=False, if_exists="append")

pandas_df_sucursal.to_csv("csv_practica/sucursal.csv", index=False)

pandas_df_sucursal.to_csv("Data/sucursal_normalizada.csv", index=False)

del pandas_df_sucursal
del sucursal

***
Vaciando el Dataframe a la Tabla `proveedor`

pandas_df_proveedor = proveedor.toPandas()

pandas_df_proveedor.to_sql(name="proveedor", con=engine, index=False, if_exists="append")

pandas_df_proveedor.to_csv("csv_practica/proveedor.csv", index=False)

del pandas_df_proveedor
del proveedor

***
Vaciando el Dataframe a la Tabla `cliente`

pandas_df_cliente = cliente.toPandas()

pandas_df_cliente.to_sql(name="cliente", con=engine, index=False, if_exists="append")

pandas_df_cliente.to_csv("csv_practica/cliente.csv", index=False)

del cliente
del pandas_df_cliente

spark.stop()
