==> CONTENEDOR-PYSPARK/05-ETL-compra.ipynb <==
#### NOTA: En este cuaderno se crearan las Tablas `compra_header` y `compra_detail` del Datawarehouse

***
### Importando Librerias

from pyspark.sql import SparkSession

from pyspark.sql.functions import count, when, isnull


***
Creando Session de Spark

spark = SparkSession.builder.appName('compra_header-compra_detail').getOrCreate()

***
Importando Conexion a Contenedor `mod4_mysql`

from db.connection import connec_db

***
Trabajando el csv asociado a  `compra`

compra = spark.read.csv("Data/Compra.csv", sep=",", header=True, inferSchema=True)

***
Visualizando como vienen los Datos

compra.show(5)

***
Visualizando la Inferencia del Schema

compra.printSchema()

***
Buscando Valores Faltantes

# Contar valores nulos por columna
compra.agg(*[count(when(isnull(c), c)).alias(c) for c in compra.columns]).show()



Creo los dataframes `compra_header` y `compra_detail` a partir de `compra`

compra_header = compra.select("IdCompra", "Fecha", "IdProveedor")

compra_detail = compra.select("IdCompra", "IdProducto", "Cantidad", "Precio")

Visualizo como se ven los nuevos Dataframes

compra_header.show()

compra_detail.show()

del compra

***
Creando la Conexion

engine = connec_db()

***
Pasando los Dataframes a Dataframes de Pandas  y vaciando en sus respectivas Tablas

# IMPORTANDO LIBRERIA PANDAS
import pandas as pd

***
Vaciando el Dataframe a la Tabla `compra_header`

pandas_df_compra_h = compra_header.toPandas()


pandas_df_compra_h.to_sql(name="compra_header", con=engine, index=False, if_exists="append")

pandas_df_compra_h.to_csv("csv_practica/compra_header.csv", index=False)

del pandas_df_compra_h
del compra_header

***
Vaciando el Dataframe a la Tabla `compra_detail`

pandas_df_compra_d = compra_detail.toPandas()

pandas_df_compra_d.to_sql(name="compra_detail", con=engine, index=False, if_exists="append")

pandas_df_compra_d.to_csv("csv_practica/compra_detail.csv", index=False)

del compra_detail
del pandas_df_compra_d

spark.stop()
