==> CONTENEDOR-PYSPARK/07-ETL-empleado-cargo-sector.ipynb <==
#### NOTA: En este cuaderno se crearan las Tablas `empleado`, `cargo` y `sector` del Datawarehouse

***
### Importando Librerias

from pyspark.sql import SparkSession

from pyspark.sql.functions import col, max, count, when, isnull, monotonically_increasing_id

***
Creando Session de Spark

spark = SparkSession.builder.appName('empleado-cargo-sector').getOrCreate()

***
Importando Conexion a Contenedor `mod4_mysql`

from db.connection import connec_db

***
Trabajando el csv asociado a  `empleado`

empleado = spark.read.csv("Data/Empleados.csv", sep=",", header=True, inferSchema=True)

***
Visualizando como vienen los Datos

empleado.show(5)

***
Visualizando la Inferencia del Schema

empleado.printSchema()

***
Buscando Valores Faltantes

# Contar valores nulos por columna
empleado.agg(*[count(when(isnull(c), c)).alias(c) for c in empleado.columns]).show()

Renombrando la columna 'ID_empleado'

empleado = empleado.withColumnRenamed("ID_empleado", "IdEmpleado")
empleado.show(1)

***
Creando el Dataframe que representa la tabla `cargo`

cargo = empleado.select("Cargo").distinct()

cargo = cargo.withColumn("IdCargo", monotonically_increasing_id()+1)

cargo.show()

***
Creando el Dataframe que representa la tabla `sector`

sector = empleado.select("Sector").distinct()

sector = sector.withColumn("IdSector", monotonically_increasing_id()+1)

sector.show()

***
Mapeando Sector en la tabla `empleado` utilizando join

map_sector = sector.withColumnRenamed("Sector", "Sector_map")

empleado_join_sector = empleado.join(map_sector, map_sector["Sector_map"] == empleado["Sector"])
empleado_join_sector.show(1)

# Verificando que el Join se Hizo correctamente Sin Nulos
empleado_join_sector.agg(*[count(when(isnull(c), c)).alias(c) for c in empleado_join_sector.columns]).show()

# Descartando la columna Sector 
empleado = empleado_join_sector.select("IdEmpleado", "Apellido", "Nombre", "Sucursal", "Cargo", "IdSector", "Salario")

del map_sector
del empleado_join_sector

***
Mapeando Cargo en la tabla `empleado` utilizando join

map_cargo = cargo.withColumnRenamed("Cargo", "Cargo_map")

empleado_join_cargo = empleado.join(map_cargo, map_cargo["Cargo_map"] == empleado["cargo"])

# Verificando que el Join se Hizo correctamente Sin Nulos
empleado_join_cargo.agg(*[count(when(isnull(c), c)).alias(c) for c in empleado_join_cargo.columns]).show()

# Descartando la columna Cargo
empleado = empleado_join_cargo.select("IdEmpleado","Apellido","Nombre","Sucursal","IdSector","IdCargo","Salario")


del map_cargo
del empleado_join_cargo

Se debe Normalizar la Columna `sucursal` para poder mapearla

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Mendoza1"]), 'Mendoza-1')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Mendoza 1"]), 'Mendoza-1')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Mendoza2"]), 'Mendoza-2')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Mendoza 2"]), 'Mendoza-2')\
                               .otherwise(col('Sucursal')))


empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["MDQ1"]), 'Mdq-1')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["MDQ2"]), 'Mdq-2')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Rosario1"]), 'Rosario-1')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Rosario2"]), 'Rosario-2')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Palermo 1"]), 'Palermo-1')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Palermo 2"]), 'Palermo-2')\
                               .otherwise(col('Sucursal')))

empleado = empleado.withColumn('Sucursal', when(col('Sucursal')\
                               .isin(["Cordoba Quiroz"]), 'Córdoba Quiroz')\
                               .otherwise(col('Sucursal')))

empleado.select("Sucursal").distinct().show(31)

***
Para poder mapear la columna 'sucursal' dentro de `empleado` necesito traer la informacion de <br>
la tabla `sucursal` ya normalizada. Cuando se trabajo 'sucursal' se dejo un archivo csv llamado <br>
'sucursal_normalizada.csv` 

sucursal = spark.read.csv("Data/sucursal_normalizada.csv", sep=",", header=True ,inferSchema=True)

sucursal.show(1)

sucursal_map = sucursal.select("IdSucursal", "Sucursal")
sucursal_map = sucursal_map.withColumnRenamed("Sucursal", "Sucur_map")

sucursal_map.show(1)

del sucursal

Mapeando Sucursal en la tabla `empleado` utilizando join

empleado_join_sucursal = empleado.join(sucursal_map, sucursal_map["Sucur_map"] == empleado["Sucursal"])

# Verificando que el Join se Hizo correctamente Sin Nulos
empleado_join_sucursal.agg(*[count(when(isnull(c), c)).alias(c) for c in empleado_join_sucursal.columns]).show()

# Descartando la Columna Sucursal
empleado = empleado_join_sucursal.select("IdEmpleado", "Apellido", "Nombre", "IdSector",
                                         "IdSucursal", "IdCargo", "Salario")

empleado.count()

del empleado_join_sucursal
del sucursal_map

***
#### Tratamiento de Valores Repetidos en `IdEmpleado`
Creo la columna "CodigoEmpleado" y le asigno el mismo valor que hasta ahora tiene "IdEmpleado"

empleado = empleado.withColumn("CodigoEmpleado", col("IdEmpleado"))
empleado.show(10)

***
En la Clase se Establecio que el nuevo `IdEmpleado` seria (sucursal * 1.000.000) + IdEmpleado_viejo

# Borro la columna `IdEmpleado`, la informacion ya la tengo en `CodigoEmpleado`
empleado = empleado.drop('IdEmpleado')

# Creo nuevamente la columna `IdEmpleado` pero esta vez (sucursal * 1.000.000) + CodigoEmpleado
empleado = empleado.withColumn("IdEmpleado", (col("IdSucursal") * 1000000) + col("CodigoEmpleado"))

empleado.show()

***
Creando la Conexion

engine = connec_db()

***
Pasando el Dataframe a Dataframe de Pandas  y vaciando en su respectiva Tabla

# IMPORTANDO LIBRERIA PANDAS
import pandas as pd

***
Vaciando el Dataframe a la Tabla `cargo`

pandas_df_cargo = cargo.toPandas()


pandas_df_cargo.to_sql(name="cargo", con=engine, index=False, if_exists="append")

pandas_df_cargo.to_csv("csv_practica/cargo.csv")

del cargo
del pandas_df_cargo

***
Vaciando el Dataframe a la Tabla `sector`

pandas_df_sector = sector.toPandas()

pandas_df_sector.to_sql(name="sector", con=engine, index=False, if_exists="append")

pandas_df_sector.to_csv("csv_practica/sector.csv", index=False)

del sector
del pandas_df_sector

***
Vaciando el Dataframe a la Tabla `empleado`

pandas_df_empleado = empleado.toPandas()

pandas_df_empleado.to_sql(name="empleado", con=engine, index=False, if_exists="append")

pandas_df_empleado.to_csv("csv_practica/empleado.csv", index=False)

del empleado

Guardo la informacion de empleado ya normalizado en un csv por si es necesario cuando se trabaje `venta`

pandas_df_empleado.to_csv("Data/empleado_normalizado.csv")

del pandas_df_empleado

spark.stop()
