==> CONTENEDOR-PYSPARK/08-ETL-venta.ipynb <==
#### NOTA: En este cuaderno se crearan las Tablas `venta_header` y `venta_detail` del Datawarehouse

***
### Importando Librerias

from pyspark.sql import SparkSession

from pyspark.sql.functions import col, count, when, isnull

***
Creando Session de Spark

spark = SparkSession.builder.appName('venta_header-venta_detail').getOrCreate()

***
Importando Conexion a Contenedor `mod4_postgres`

from db.connection import connec_db

***
Trabajando el csv asociado a  `compra`

venta = spark.read.csv("Data/Venta.csv", sep=",", header=True, inferSchema=True)

***
Visualizando como vienen los Datos

venta.show(5)

***
Visualizando la Inferencia del Schema

venta.printSchema()

***
Buscando Valores Faltantes

# Contar valores nulos por columna
venta.agg(*[count(when(isnull(c), c)).alias(c) for c in venta.columns]).show()



En las cantidades Faltantes segun los Homeworks se asumio colocar 1 <br>
En cuanto a los precios se localizaron todos en la tabla `producto` y se buscan por 'IdProducto`


# Se asigna 1 a las cantidades
venta = venta.fillna({'Cantidad':1})

# Se coloca 0 por ahora a los precios
venta = venta.fillna({'Precio':0})

***
Hay que obtener la informacion de los precios de la data de Productos

producto = spark.read.csv("Data/PRODUCTOS.csv", sep=",", header=True, inferSchema=True)

producto = producto.withColumnRenamed("Precio","precio_p")

producto.show(1)

venta.count()

venta_join_prod = venta.join(producto, producto["ID_PRODUCTO"] == venta["IdProducto"], "left")

venta_join_prod.count()

venta_join_prod.show(1)

venta_join_prod = venta_join_prod.withColumn(
    "PrecioFinal",
    when(venta_join_prod['Precio'] == 0, col("precio_p"))
    .otherwise(col('Precio'))
)

venta_join_prod.show(1)

# SELECCIONO LAS COLUMNAS 
venta = venta_join_prod.select("IdVenta", "Fecha", "Fecha_Entrega", "IdCanal", "IdCliente",
                               "IdSucursal","IdEmpleado","IdProducto","Cantidad","PrecioFinal")

del venta_join_prod

*** 
Corrijo el codigo de Empleado en una nueva columna cuyo nombre sera tempEmpleado

venta = venta.withColumn("tempEmpleado", (col("IdSucursal") * 1000000) + col("IdEmpleado"))

Borro la columna 'IdEmpleado'

venta = venta.drop("IdEmpleado")

Renombro Columnas

venta = venta.withColumnRenamed("tempEmpleado", "IdEmpleado")\
             .withColumnRenamed("PrecioFinal", "Precio")\
             .withColumnRenamed("Fecha_Entrega", "FechaEntrega")

***
Creo los dataframe `venta_header` y `venta_detail` a partir de `venta`

venta_header = venta.select("IdVenta", "Fecha", "FechaEntrega", "IdCanal", "IdCliente", "IdSucursal", "IdEmpleado")

venta_header.show(1)

venta_detail = venta.select("IdVenta", "IdProducto", "Precio", "Cantidad")

venta_detail.show(1)

del venta

***
Creando la Conexion

engine = connec_db()

***
Pasando los Dataframes a Dataframes de Pandas  y vaciando en sus respectivas Tablas

# IMPORTANDO LIBRERIA PANDAS
import pandas as pd

***
Vaciando el Dataframe a la Tabla `venta_header`

pandas_df_venta_h = venta_header.toPandas()


pandas_df_venta_h.to_sql(name="venta_header", con=engine, index=False, if_exists="append")

pandas_df_venta_h.to_csv("csv_practica/venta_header.csv", index=False)

del pandas_df_venta_h
del venta_header

***
Vaciando el Dataframe a la Tabla `venta_detail`

pandas_df_venta_d = venta_detail.toPandas()

pandas_df_venta_d.to_sql(name="venta_detail", con=engine, index=False, if_exists="append")

pandas_df_venta_d.to_csv("csv_practica/venta_detail.csv", index=False)

del venta_detail
del pandas_df_venta_d

spark.stop()
